Renames processor file.
