+++ ts/common/processor.ts
